// src/app/model/data-component.ts
export interface DataComponent {
  activeInd?: boolean;
  componentType?: string;
  dataMappingName?: string;
  dataValueEditableInd?: boolean;
  dataValueType?: string;
  description?: string;
  displayOrderNumber?: number;
  filterableInd?: boolean;
  filterListOptionInd?: boolean;
  name?: string;
  sectionName?: string;
  shortName?: string;
  sortableInd?: boolean;
  value?: string;
  paramDropdownValues?: string[];
}